# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText
  description: |
    This API allows users to interact with each other 
    on the WasaText app.
    Users can create groups, private chats, send text or image messages, 
    reply, forward or react them.
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: |
        Use the user identifier returned by `/login` in the Authorization header.
        Example: `Authorization: Bearer abcdef012345`
  responses:
    UnauthorizedError:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
...
security:
  - BearerAuth: []
paths:
...
  /groups/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        description: The group id
        schema:
          type: integer
    post:
      tags:
        - Groups
      summary: Add a new member to a group
      description: Any user belonging to a group can add other users to the group if they are not already there.
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The username of the member to add
              type: object
              required:
                - userId
              properties:
                userId:
                  type: integer
                  description: user id
      responses:
        "200":
          description: Member added to the group successfully.
          content:
            application/json:
              schema:
                type: object
                description: A descriptive message of success
                example:
                  message: Member added to the group successfully

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: Forbidden
              example:
                message: Forbidden
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                description: A descriptive message of different failures
                type: object
                properties:
                  message:
                    type: string
                    description: >
                      error:
                      * `First Error` - User not found 
                      * `Second Error` - Group not found
                    enum:
                      - user_not_found
                      - group_not_found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      tags:
        - Groups
      summary: Leave a group
      description: A user can remove themselves from a group they belong to, or if he is admin, remove another user. If the user who deletes himself is an admin, another user will be promoted to admi will be promoted to admin.
      operationId: leaveGroup
      responses:
        "204":
          description: User successfully removed from the group
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: User not part of the group
              example:
                message: User not part of the group
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: Group not found
              example:
                message: Group not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /groups:
    post:
      tags:
        - Groups
      summary: Create group
      description: Create a new group with a name and members.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Fields to create a group with
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 30
                  description: The name of the group
                members:
                  type: array
                  description: List of user ids to add to the group
                  items:
                    type: integer
            example:
              name: I ragazzi del calcetto
              members: [2, 6, 8]
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                description: A descriptive message of success and fields of the newly created group
                type: object
                properties:
                  groupId:
                    type: integer
                    description: The group id
                  name:
                    type: string
                    description: The group name
        "401":
          $ref: "#/components/responses/UnauthorizedError"